requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
feedparser==6.0.11
gigachat==0.1.37
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.29
langchain-gigachat==0.3.2
langsmith==0.2.10
orjson==3.10.14
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
sniffio==1.3.1
tenacity==9.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
